# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {
      "name": "API Support",
      "url": "https://github.com/fransfilastap/kontak"
    },
    "title": "Kontak API",
    "version": "1.0",
    "description": "WhatsApp-to-REST-API bridge",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths": {
...
    "/admin/groups/{client_id}": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get all joined groups for a device",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "groups"
        ],
        "summary": "List groups",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/groups/{client_id}/sync": {
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Fetch and sync joined groups from WhatsApp for a device",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "groups"
        ],
        "summary": "Sync groups",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "pkg_http.ErrorResponse": {
      "type": "object",
      "properties": {
        "context": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
  },
  "host": "localhost:8080",
  "basePath": "/"
}